requirements.txt
pyknp==0.6.1
PyYAML==6.0.1
six==1.16.0
tqdm==4.66.4
